include/net/netfilter/nf_conntrack.h: decrease memory usage from 224 to 216 bytes per 'struct nf_conn'

+++ include/net/netfilter/nf_conntrack.h
@@ -145,8 +145,8 @@ struct nf_conn {
 
 #if IS_ENABLED(CONFIG_FAST_NAT)
 	/* fast nat ext */
-	u_int32_t fast_ext;
-	u_int32_t fast_nat_binded;
+	u_int8_t fast_ext;
+	u_int8_t fast_nat_binded;
 #endif
 
 	/* Extensions */
